+++ reductstore/src/api/http.rs
@@ -869,7 +869,7 @@ pub(crate) mod tests {
                     dst_host: "http://localhost:8080".to_string(),
                     dst_token: None,
                     entries: vec![],
-                    prefix: String::new(),
+                    dst_prefix: String::new(),
                     include: Default::default(),
                     exclude: Default::default(),
                     each_n: None,