+++ reductstore/src/cfg/provision/replication.rs
@@ -65,7 +65,7 @@ impl<EnvGetter: GetEnv, ExtCfg: ExtCfgBounds> CfgParser<EnvGetter, ExtCfg> {
                     dst_host: "http://localhost".to_string(),
                     dst_token: None,
                     entries: vec![],
-                    prefix: String::new(),
+                    dst_prefix: String::new(),
                     include: Labels::default(),
                     exclude: Labels::default(),
                     each_n: None,
@@ -130,10 +130,10 @@ impl<EnvGetter: GetEnv, ExtCfg: ExtCfgBounds> CfgParser<EnvGetter, ExtCfg> {
                 replication.settings.entries = entries.split(",").map(|s| s.to_string()).collect();
             }
 
-            if let Some(prefix) =
-                env.get_optional::<String>(&format!("RS_REPLICATION_{}_PREFIX", id))
+            if let Some(dst_prefix) =
+                env.get_optional::<String>(&format!("RS_REPLICATION_{}_DST_PREFIX", id))
             {
-                replication.settings.prefix = prefix;
+                replication.settings.dst_prefix = dst_prefix;
             }
 
             for (key, value) in env.matches(&format!("RS_REPLICATION_{}_INCLUDE_(.*)", id)) {
@@ -253,7 +253,7 @@ mod tests {
         assert_eq!(replication.dst_host, "http://localhost/");
         assert_eq!(replication.dst_token, Some("TOKEN".to_string()));
         assert_eq!(replication.entries, vec!["entry1", "entry2"]);
-        assert_eq!(replication.prefix, "robot-1");
+        assert_eq!(replication.dst_prefix, "robot-1");
         assert_eq!(replication.each_n, Some(10));
         assert_eq!(replication.each_s, Some(0.5));
         assert_eq!(
@@ -528,7 +528,7 @@ mod tests {
                 dst_host: "http://localhost".to_string(),
                 dst_token: None,
                 entries: vec![],
-                prefix: String::new(),
+                dst_prefix: String::new(),
                 include: Labels::default(),
                 exclude: Labels::default(),
                 each_n: None,
@@ -618,7 +618,7 @@ mod tests {
                 dst_host: "http://localhost".to_string(),
                 dst_token: None,
                 entries: vec![],
-                prefix: String::new(),
+                dst_prefix: String::new(),
                 include: Labels::default(),
                 exclude: Labels::default(),
                 each_n: None,
@@ -703,7 +703,7 @@ mod tests {
                 dst_host: "http://localhost".to_string(),
                 dst_token: None,
                 entries: vec![],
-                prefix: String::new(),
+                dst_prefix: String::new(),
                 include: Labels::default(),
                 exclude: Labels::default(),
                 each_n: None,
@@ -803,7 +803,7 @@ mod tests {
             .return_const(Ok("entry1,entry2".to_string()));
         mock_getter
             .expect_get()
-            .with(eq("RS_REPLICATION_1_PREFIX"))
+            .with(eq("RS_REPLICATION_1_DST_PREFIX"))
             .return_const(Ok("robot-1".to_string()));
 
         mock_getter